hash: Make HSH_Fail() rush the waiting list
This is not a nice caching outcome. In fact, when a fetch fails we don't
know the outcome. In that case we need to rush one request that will
trigger a new fetch and hopefully complete with a proper outcome to rush
other requests accordingly.

As far as lookup is concerned there is no point in keeping the OC_F_BUSY
flag so in the case where a failure happens before a proper HSH_Unbusy()
it can be dropped before the rush. This would otherwise happen once the
last objcore reference is dropped. One is held by the current fetch task
and the other one by the client task that triggerred the fetch. It would
amount to an unnecessary delay.

The OC_F_FAILED flag will prevent it from being looked up, so there is
no reason to delay the rush until the final objcore reference is dropped.

This also means that once the client task that triggered the fetch task
resumes it can only be operating on a non-busy object.