vbe: Specialize pipe mode for CONNECT transactions
The main difference between pipe and CONNECT is that only pipe sends a
bereq and needs to create a task for the busyobj. On the other hand we
will send a resp in CONNECT mode. Since CONNECT mode establishes a
tunnel, this must force a fresh connection. This cannot be used like
pipe mode with HTTP upgrades.

CONNECT transactions get their dedicated set of timestamps.