waiter: report WAITER_REMCLOSE correctly
There seems to be a discrepancy between the epoll and kqueue
implementation. One implementation is not reporting WAITER_REMCLOSE,
while the other implementation is reporting WAITER_REMCLOSE too soon.

epoll:
We sometimes incorrectly report WAITER_ACTION when it is supposed to be
WAITER_REMCLOSE. We can check EPOLLRDHUP before reading 1 byte with MSG_PEEK to
detect WAITER_REMCLOSE. This saves us from a pontential session detour from the
Waiter to a worker thread.

kqueue:
We incorrectly report WAITER_REMCLOSE when there could be more data to be
read from the socket. Change this so that we perform read of 1 byte with
MSG_PEEK to detect closure when EV_EOF is set.

DESCRIPTION:EVFILT_READ:Sockets:
"It is possible for EOF to be  returned (indicating the connection is gone)
while there is still data pending in the socket buffer."
- kqueue (2)

Signed-off-by: Asad Sajjad Ahmed <asadsa@varnish-software.com>